replaced net code with rpc code
