--- app/app2/main.go
@@ -1,18 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"github.com/aadit-n3rdy/hotCrossBuns/replica"
-)
-
-func main() {
-	rep := new(replica.Replica)
-
-	rep.New(2)
-	rep.Start()
-
-	for _, rep := range rep.ReturnOtherReplicas() {
-		fmt.Println(rep)
-	}
-}
